# src/main/resources/static/swagger-ui/openapi3.yaml
openapi: 3.0.1
info:
  title: Classic Daramg's API docs
  description: 클래식 듣는 다람쥐의 API 문서입니다.
  version: 0.1.0
servers:
- url: http://localhost:8080
...
paths:
  /auth/email-verifications:
    post:
      tags:
      - Auth API
      summary: 비밀번호 초기화 시 인증코드 이메일 발송
      description: 비밀번호 초기화를 위해 이메일 주소로 인증코드를 발송합니다.
      operationId: 비밀번호_초기화를_위한_인증코드_메일_발송회원가입을_위한_인증코드_메일_발송
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/auth-email-verifications1086464701"
            examples:
              비밀번호_초기화를_위한_인증코드_메일_발송:
                value: "{\"email\":\"daramg123@gmail.com\",\"emailPurpose\":\"PASSWORD_RESET\"\
                  }"
              회원가입을_위한_인증코드_메일_발송:
                value: "{\"email\":\"daramg123@gmail.com\",\"emailPurpose\":\"SIGNUP\"\
                  }"
      responses:
        "200":
          description: "200"
  /auth/login:
    post:
      tags:
      - Auth API
      summary: 로그인
      description: 유저가 로그인합니다.
      operationId: 로그인
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/auth-login684347283"
            examples:
              로그인:
                value: "{\"email\":\"hamster@gmail.com\",\"password\":\"Hoshi0615!\"\
                  }"
      responses:
        "200":
          description: "200"
...
  /auth/password-reset:
    put:
      tags:
      - Auth API
      summary: 비밀번호를 잊었을 경우 비밀번호 재설정
      description: 사용자의 비밀번호를 새로운 비밀번호로 변경하고 자동으로 로그아웃 처리합니다.
      operationId: 비밀번호_재설정
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/auth-password-reset-1301565020"
            examples:
              비밀번호_재설정:
                value: "{\"email\":\"svt@pledis.com\",\"password\":\"NewPassword123!\"\
                  }"
      responses:
        "200":
          description: "200"
...
  /auth/signup:
    post:
      tags:
      - Auth API
      summary: 회원가입
      description: 새로운 유저를 생성합니다.
      operationId: 회원가입
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/auth-signup-225747861"
            examples:
              회원가입:
                value: "{\"name\":\"권순영\",\"birthdate\":\"1996-06-15\",\"email\":\"\
                  hamster@gmail.com\",\"password\":\"Password123!\",\"profileImage\"\
                  :\"https://example.com/profile.jpg\",\"nickname\":\"햄쥑이\",\"bio\"\
                  :\"나라 지키는 중\"}"
      responses:
        "201":
          description: "201"
  /auth/verify-email:
    post:
      tags:
      - Auth API
      summary: 인증번호로 이메일 인증
      description: 사용자가 제공한 인증번호로 이메일 주소를 인증합니다.
      operationId: 인증번호로_이메일_인증
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/auth-verify-email-1072797477"
            examples:
              인증번호로_이메일_인증:
                value: "{\"email\":\"daramg123@gmail.com\",\"verificationCode\":\"\
                  123456\"}"
      responses:
        "200":
          description: "200"
...
components:
  schemas:
    auth-signup-225747861:
      type: object
      properties:
        password:
          type: string
          description: "비밀번호 (영어 대/소문자, 숫자, 특수문자 포함 10자 이상)"
        birthdate:
          type: string
          description: 생년월일 (YYYY-MM-DD)
        nickname:
          type: string
          description: 닉네임 (2~8자)
        name:
          type: string
          description: 이름
        bio:
          type: string
          description: bio (12자 이하)
        profileImage:
          type: string
          description: 프로필 이미지 URL
        email:
          type: string
          description: 이메일
...
    auth-password-reset-1301565020:
      type: object
      properties:
        password:
          type: string
          description: "새 비밀번호 (영어 대/소문자, 숫자, 특수문자 포함 10자 이상)"
        email:
          type: string
          description: 변경하고자 하는 유저의 이메일
...
    auth-verify-email-1072797477:
      type: object
      properties:
        email:
          type: string
          description: 이메일
        verificationCode:
          type: string
          description: "인증번호: 6자리 숫자"
...
    auth-login684347283:
      type: object
      properties:
        password:
          type: string
          description: "비밀번호 (영어 대/소문자, 숫자, 특수문자 포함 10자 이상)"
        email:
          type: string
          description: 이메일
...
    auth-email-verifications1086464701:
      type: object
      properties:
        emailPurpose:
          type: string
          description: "이메일 인증코드 전송 목적(SIGNUP, PASSWORD_RESET)"
        email:
          type: string
          description: 이메일